refactor: extract VibeAcpClient into chainlit-free acp.py
Moves the JSON-RPC wire helpers and VibeAcpClient out of app.py so a
second frontend (the upcoming workspace server) can reuse the subprocess
client without importing Chainlit. The audit-logger handler setup moves
with it; cwd/vibe_home become constructor parameters.
